requirements.txt
streamlit
transformers
torch
pillow
gTTS
accelerate
sentencepiece